+++ pkg/traceql/engine.go
@@ -272,7 +272,10 @@ func (e *Engine) asTraceSearchMetadata(spanset *Spanset) *tempopb.TraceSearchMet
 				attribute.Intrinsic == IntrinsicTraceRootSpan ||
 				attribute.Intrinsic == IntrinsicTraceID ||
 				attribute.Intrinsic == IntrinsicSpanID ||
-				attribute.Intrinsic == IntrinsicEventName {
+				attribute.Intrinsic == IntrinsicEventName ||
+				attribute.Intrinsic == IntrinsicLinkTraceID ||
+				attribute.Intrinsic == IntrinsicLinkSpanID {
+
 				continue
 			}